artex_agent/src/livekit_rtc_stubs/livekit_rtc_pb2_grpc.py
# Placeholder for livekit_rtc_pb2_grpc.py
# Enhanced to align with placeholder livekit_rtc_pb2.py
import grpc
from . import livekit_rtc_pb2 as rtc_pb2 # Use relative import

class RTCServiceStub:
    def __init__(self, channel: grpc.aio.Channel):
        """Constructor for RTCServiceStub.

        Args:
            channel: A grpc.aio.Channel object.
        """
        self.channel = channel

        # Define the bi-directional Signal stream method
        # The path '/livekit.RTCService/Signal' must match the service FQN in the .proto
        self.Signal = channel.stream_stream(
                '/livekit.RTCService/Signal', # Standard FQN for LiveKit RTC service
                request_serializer=rtc_pb2.SignalRequest.SerializeToString,
                response_deserializer=rtc_pb2.SignalResponse.FromString,
        )

        # Example of other potential (unary) methods if they existed on RTCService directly:
        # self.SendDTMF = channel.unary_unary(
        #         '/livekit.RTCService/SendDTMF',
        #         request_serializer=rtc_pb2.SendDTMFRequest.SerializeToString,
        #         response_deserializer=rtc_pb2.SendDTMFResponse.FromString,
        # )

class RTCServiceServicer:
    """Placeholder for server-side implementation of RTCService."""
    def Signal(self, request_iterator, context: grpc.aio.ServicerContext):
        # This method would handle incoming signals from a client and send responses.
        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
        context.set_details('Method not implemented!')
        raise NotImplementedError('Method not implemented!')

    # Add other methods from the service definition here if needed for the servicer.

def add_RTCServiceServicer_to_server(servicer: RTCServiceServicer, server: grpc.Server):
    """Registers the servicer with a gRPC server. Not used by the client."""
    # This would typically involve:
    # import livekit_rtc_pb2 # Ensure this is the correct import path if used here
    # rpc_method_handlers = {
    #     'Signal': grpc.stream_stream_rpc_method_handler(
    #         servicer.Signal,
    #         request_deserializer=livekit_rtc_pb2.SignalRequest.FromString, # Use correct pb2 import
    #         response_serializer=livekit_rtc_pb2.SignalResponse.SerializeToString,
    #     ),
    # }
    # generic_handler = grpc.method_handlers_generic_handler(
    #     'livekit.RTCService', rpc_method_handlers) # Use correct FQN
    # server.add_generic_rpc_handlers((generic_handler,))
    pass # Placeholder for now

print("Placeholder livekit_rtc_pb2_grpc.py loaded (enhanced).")
